core: fix variable sizes and initializations

seen by compile warnings on FreeBSD with clang++

+++ core/sip/pcap_logger.cpp
@@ -125,8 +125,8 @@ int pcap_logger::log(const char *data, int data_len,
   hdr.ip.ip_hl = 5;
   hdr.ip.ip_v = 4;
   hdr.ip.ip_tos = 0;
-  hdr.ip.ip_len = htons(size);
-  hdr.ip.ip_id = htonl(54321);
+  hdr.ip.ip_len = (unsigned short) htons(size);
+  hdr.ip.ip_id = (unsigned short)htonl(54321);
   hdr.ip.ip_off = 0;
   hdr.ip.ip_ttl = 255;
   hdr.ip.ip_p = 0x11; // UDP
